rename js react files to jsx extensions
